src/pyproject.toml
[tool.poetry]
name = "kafka-study"
version = "0.1.0"
description = ""
authors = ["freemjstudio <mjwoo0@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"
msgpack = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
